refactor: created shell for datamodel
