+++ cognite/powerops/resync/models/cogshop.py
@@ -1,0 +1,5 @@
+from ._base import DataModel
+
+
+class CogShopModel(DataModel):
+    ...
