+++ cognite/powerops/resync/models/core.py
@@ -6,7 +6,7 @@ from cognite.client.data_classes import Asset, TimeSeries
 from pydantic import Field
 
 from cognite.powerops.cdf_labels import AssetLabel
-from cognite.powerops.resync.models.base import AssetModel, AssetType, CDFSequence
+from cognite.powerops.resync.models._base import AssetModel, AssetType, CDFSequence
 
 
 class Generator(AssetType):
