+++ cognite/powerops/resync/models/market/rkom.py
@@ -5,7 +5,7 @@ from typing import ClassVar
 from pydantic import Field
 
 from cognite.powerops.cdf_labels import AssetLabel
-from cognite.powerops.resync.models.base import AssetType, CDFSequence, NonAssetType
+from cognite.powerops.resync.models._base import AssetType, CDFSequence, NonAssetType
 from cognite.powerops.resync.models.market.base import Bid, Market, Process
