+++ cognite/powerops/resync/to_models/to_core_model.py
@@ -7,7 +7,7 @@ from typing import Optional
 import pandas as pd
 from cognite.client.data_classes import Sequence, TimeSeries
 
-import cognite.powerops.resync.models.base
+import cognite.powerops.resync.models._base
 from cognite.powerops.resync.config_classes.core.connections import Connection
 from cognite.powerops.resync.config_classes.resync_config import CoreConfigs
 from cognite.powerops.resync.models import core
